requirements.txt
cmake==3.22.4
dlib==19.23.1
fastapi==0.75.2
imutils==0.5.4
mysql-connector-python==8.0.25
notebook==6.4.0
numpy==1.21.0
opencv-python==4.5.4.58
pandas==1.3.1
Pillow==8.3.1
virtualenv==20.6.0
